+++ src/app/app-routing.module.ts
@@ -1,3 +1,4 @@
+import { OfferListPaginatedComponent } from './sites/offers/pages/offer-list-paginated/offer-list-paginated.component';
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
@@ -25,6 +26,10 @@ const routes: Routes = [
   },
   {
     path: 'kurse',
+    component: OfferListPaginatedComponent,
+  },
+  {
+    path: 'allekurse',
     component: OfferListComponent,
   },
   {