+++ src/app/sites/ki-tools/components/shared/templates/page_loader.ts
@@ -20,64 +20,20 @@ export class PageLoader implements OnInit {
     loadingStatus: KIToolsTypes.ScriptLoadingStatus;
     kitoolsAreOnline: boolean;
     lnkKITools = this.staticService.getPathInfo().lnkKITools;
-  
+
     constructor(
-      public renderer: Renderer2,
       public staticService: StaticService,
-      public kiStatusService: KiStatusService
     ) {}
-  
-    ngOnInit(loadPackage: boolean = false): void {
-      loadPackage = false; // Die werden nicht mehr geladen
+
+    ngOnInit(): void {
       this.loadingStatus = {
-        isLoaded: false,
+        isLoaded: true,
         isError: false,
-      };
-      
-      if (loadPackage){
-        this.kitoolsAreOnline = this.staticService.getKIConfig().online;
-        if (this.kitoolsAreOnline) {
-          console.log('Laden der KI Packages');
-          this.loadKIPackages();
-        }
-      } else {
-        this.loadingStatus = {
-          isLoaded: true,
-          isError: false,
-        };
       }
     }
 
     onModalClose() {
      // Kein modales Fenster
-    }
-  
-    // Falls die Scripte nicht geladen sind
-    loadKIPackages() {
-      this.kiStatusService.loadKIScript(this.renderer).subscribe(
-        (scripts) => {
-          if (!KIToolsHelper.checkLoadedScripts(scripts)) {
-            // Eines der Scripte wurde nicht geladen
-            console.log('Scriptfehler');
-            this.loadingStatus = {
-              isLoaded: false,
-              isError: true,
-            };
-          } else {
-            this.loadingStatus = {
-              isLoaded: true,
-              isError: false,
-            };
-          }
-        },
-        (error) => {
-          console.log('Error: ', error);
-          this.loadingStatus = {
-            isLoaded: false,
-            isError: true,
-          };
-        }
-      );
     }
 
 }