+++ src/app/sites/ki-tools/pages/sentiment/sentiment-modal.component.ts
@@ -31,7 +31,6 @@ export class NgbdSentimentModalComponent implements OnInit {
   constructor(public activeModal: NgbActiveModal) {}
 
   ngOnInit(): void {
-    // Im modalen Fenster sollte alles geladen sein.
     this.loadingStatus = {
       isLoaded: true,
       isError: false,