+++ src/app/sites/offers/components/filter-row/filter-row.directive.ts
@@ -6,13 +6,8 @@ import {
   Output,
   EventEmitter,
 } from '@angular/core';
-import { FilterComboBoxComponent } from './filter-combo-box/filter-combo-box.component';
+import { FilterComboBoxComponent, FilterFunctionCallbackItem } from './filter-combo-box/filter-combo-box.component';
 import { Offer } from 'src/app/core/models/offer';
-
-export type FilterFunctionCallbackItem = {
-  type: string,
-  id: number
-}
 
 @Directive({ selector: 'app-filter-row, [app-filter-row]' })
 export class FilterRowDirective implements AfterContentInit {